The final offensive of the Spanish Civil War took place between 26 March and 1 April 1939, towards the end of the Spanish Civil War. On 5 March 1939, the Republican Army, led by Colonel Segismundo Casado and the politician Julián Besteiro, rose against the socialist prime minister Juan Negrín, and formed a military junta, the National Defence Council (Consejo Nacional de Defensa or CND) to negotiate a peace deal. Negrín fled to France but the communist troops around Madrid rose against the junta, starting a civil war within the civil war. Casado defeated them and started peace negotiations with the Nationalists. Francisco Franco, however, was willing to accept only an unconditional surrender. On 26 March, the Nationalists started a general offensive and by 31 March, they controlled all of Spanish territory. Hundreds of thousands of Republicans were arrested and interned in concentration camps.
Background
Fall of Catalonia
After the fall of Catalonia in February 1939, the military situation of the Republic was hopeless. Despite still having the capital city and approximately 30% of Spanish territory, it had lost 220,000 soldiers, the second most populated city in the country and the industrial resources of Catalonia.[9] Furthermore, on 27 February, President Manuel Azaña resigned. The United Kingdom and France then recognised the Nationalist government.[10]

Military situation
The Republican army still had between 250,000[11] and 500,000 men[12] but only 40 aircraft (three Natasha and two Katiuska bomber squadrons, and 25 Chatos and Moscas fighters),[13] little artillery and few automatic weapons.[14] Many soldiers were unarmed (in December 1938, the Republican army had only 225,000 rifles),[15] and lacked shoes and overcoats.[16] In Madrid, there was food for only two months and no water, heating, medicine or surgical dressings.[17] On the other hand, the Nationalist army had more than a million men at the end of 1938, with 35,000 Moroccans, 32,000 Italians and 5,000 Germans,[18] as well as 600 aircraft.[19]

Opposition to continued resistance
On 16 February, the high command of the Republican Army told Prime Minister Juan Negrín that further military resistance was impossible.[20] Most of the members of the Republican Army, the PSOE, the UGT and the CNT believed that it was necessary to initiate peace negotiations.[21] Nevertheless, Negrín, backed by the communist PCE, wanted to continue fighting because Franco rejected giving any guarantee against reprisals and a continental war against fascism was believed to be imminent.[22] Furthermore, Negrín wanted to organise the evacuation of those who were most at risk.[23]