Badlands are a type of dry terrain where softer sedimentary rocks and clay-rich soils have been extensively eroded.[1] They are characterized by steep slopes, minimal vegetation, lack of a substantial regolith, and high drainage density.[2] Ravines, gullies, buttes, hoodoos and other such geologic forms are common in badlands.
Badlands are found on every continent except Antarctica, being most common where there are unconsolidated sediments. They are often difficult to navigate by foot, and are unsuitable for agriculture. Most are a result of natural processes, but destruction of vegetation by overgrazing or pollution can produce anthropogenic badlands.

Badlands topography
Badlands are characterized by a distinctive badlands topography.[3][4] This is terrain in which water erosion has cut a very large number of deep drainage channels, separated by short, steep ridges (interfluves).[5] Such a drainage system is said to have a very fine drainage texture,[6] as measured by its drainage density. Drainage density is defined as the total length of drainage channels per unit area of land surface. Badlands have a very high drainage density of 48 to 464 kilometres per square kilometre (77 to 747 miles per square mile).[5] The numerous deep drainage channels and high interfluves creates a stark landscape of hills, gullies, and ravines.[3]

Caprock on hoodoos in Theodore Roosevelt National Park
In addition to a dense system of drainages and interfluves, badlands often contain buttes and hoodoos. These are formed by resistant beds of sandstone, which form the caprock of the buttes and hoodoos.[4]

Origin
Badlands arise from a combination of an impermeable but easily eroded ground surface, sparse vegetation, and infrequent but heavy rainfall.[6] The surface bedrock is typically mudrock, sometimes with evaporites, with only occasional beds of more resistant sandstone.[5] Infrequent heavy rains lead to heavy erosional dissection. Where sudden precipitation cannot penetrate impermeable clays, it is channeled into a very dense system of streamlets that erode a dense system of ever-enlarging, coalescing gulleys and ravines. Erosion is enhanced by pelting raindrops that dislodge soft sediments. The presence of bentonite clay further increases erodability, as can rejuvenation of the drainage system from regional uplift, as occurred at Badlands National Park.[4]

In addition to surface erosion, badlands sometimes have well-developed piping, which is a system of pipes, joints, caverns, and other connected void spaces in the subsurface through which water can drain. However, this is not a universal feature of badlands. For example, the Henry Mountains badlands show very little piping.[2]

The precise processes by which the erosion responses take place vary depending on the precise interbedding of the sedimentary material.[7] However, it has been estimated that the badlands of Badlands National Park erode at the relatively high rate of about one inch or 25 millimetres per year.[8] The White River draining Badlands National Park was so named for its heavy load of bentonite clay eroded from the badlands.[4]

Regolith
Badlands are partially characterized by their thin to nonexistent regolith layers. The regolith profiles of badlands in arid climates are likely to resemble one another. In these regions, the upper layer (~1–5 cm or 0.4–2.0 in) is typically composed of silt, shale, and sand (a byproduct of the weathered shale). This layer can form either a compact crust or a looser, more irregular aggregation of "popcorn" fragments. Located beneath the top layer is a sublayer (~5–10 cm or 2.0–3.9 in), below which can be found a transitional shard layer (~10–40 cm or 3.9–15.7 in), formed largely of loose disaggregated shale chips, which in turn eventually gives way to a layer of unweathered shale. Badlands such as those found in the Mancos Shale, the Brule Formation, the Chadron Formation, and the Dinosaur Provincial Park can be generally said to fit this profile.[2]

In less arid regions, the regolith profile can vary considerably. Some badlands have no regolith layer whatsoever, capping instead in bare rock such as sandstone. Others have a regolith with a clay veneer, and still others have a biological crust of algae or lichens.[2]

In addition to lacking significant regolith, they also lack much vegetation. The lack of vegetation could very well be a result of the lack of a substantial regolith.[9]