Okinawa soba (沖縄そば) is a type of noodle produced in Okinawa Prefecture, Japan. Okinawa soba is a regional collective trademark of The Okinawa Noodle Manufacturing Co-op.[1] On Okinawa, it is sometimes simply called soba (or suba in Okinawan dialects), although this Japanese term typically refers to buckwheat noodles in mainland Japan.
The thick wheat noodles more closely resemble the texture of udon, and when served in soup, the broth is more similar to that of ramen. The noodles tend to have a circular cross section in the Yaeyama Islands and tend to be slightly flat in the rest of Okinawa Prefecture.

It is served in a broth flavored with konbu (edible seaweed), katsuobushi flakes and pork. Standard toppings are kamaboko (fish cake), sliced scallion and a thick slice of stewed san-mai niku (三枚肉, pork belly; lit. "three-layer meat") or soki (boneless pork ribs) and usually garnished with beni shōga (pickled ginger). For extra spice, diners can add a few drops of kōrēgūsu, which consists of chile peppers soaked in awamori rice liquor.

Varieties include:

Sōki soba (ソーキそば) - topped with extra ribs, known as sōki in Okinawan.
Tebichi soba (てびちそば) - topped with stewed pig's trotters.
Name
Today the word soba means buckwheat, in particular buckwheat noodles, in Japanese. This word is in origin an abbreviation of soba-mugi (buckwheat). The word soba refers to the edge in ancient Japanese and soba-mugi meant "edgy wheat". Around the 16th century, buckwheat began to be processed into thin noodles. This method was originally known as soba-kiri (lit. soba cutting) but it was later abbreviated into soba. Soba was mainly consumed in eastern Japan and was not popular in western Japan or on Okinawa. Buckwheat noodles remain rare on Okinawa, where they are known as Nihon soba (Japanese soba) or kuroi soba (black soba).

In modern Japan, the meaning of soba was extended so that it could refer to other types of thin noodles. Yakisoba (stir-fried noodles) contain no buckwheat. Shina soba, commonly known as Chūka soba (Chinese soba), or rāmen today, is also made from wheat flour. Okinawa soba falls into this category and is made entirely from wheat.

Under a fair competition regulation, soba was standardized as "the noodles contain at least 30% of buckwheat." This definition caused a problem when Okinawa was returned to Japanese rule in 1972. In 1976, the Fair Trade Commission tried to make Okinawa Prefecture drop the name "soba" because Okinawa's soba did not conform to the regulation. The Okinawa Noodle Manufacturing Co-op negotiated with the Fair Trade Commission and as a result it was accepted as one of several exceptions under the name of Okinawa soba.

History
Some attempt to trace the origin of Okinawa soba to the Ryūkyū Kingdom period. Chinese delegates possibly brought wheat noodles to the royal court,[2] though no link has been established to the modern Okinawan cuisine.[3]

The first known documented reference to Okinawa soba was in 1902 during the late Meiji period when it was known as Shina soba (Chinese soba) as in mainland Japan. Several soba restaurants are known to have operated in Naha during the Taishō period and to have close links with the red-light district.[3] Soba restaurants were destroyed by U.S. attacks during the Battle of Okinawa. Its revival began in U.S. internment camps, where wheat rations were supplied. In the 1960s, soba was popularized with the introduction of noodle making machines. The popularization was accompanied by the transformation from a restaurant meal to home cooking. The mainland Japanese custom of eating soba on New Year's Eve was introduced to Okinawa in 1968, but buckwheat noodles were replaced by wheat noodles. Today, Okinawa soba is considered as a vital part of traditional Okinawan culture, although folklorist Nishimura Hidemi argued that it was an invented tradition.[3]

The soba has been adopted as a local tradition in Campo Grande (in the state of Mato Grosso do Sul, Brazil), due to the large Okinawan-descended population in the city of 850,000. It's usual to find specialized restaurants (sobarias) along the city serving a local adaptation of the dish known as sobá due to the difficulty of finding seafish in that region of Brazil.