Yahoo! Japan (ヤフー, Yafū) is a Japanese web portal. Its search engine was the most-visited website in Japan, nearing monopolistic status.[1]
According to The Japan Times, as of 2012, Yahoo Japan had a footprint on the internet market in Japan. In terms of use as a search engine, however, it has never surpassed Google. The company is the second largest search engine used in Japan as of July 2021, with a market share of 19% behind Google's 77%.[2]

Yahoo! Japan search engine was a directory-type search engine, similar to Yahoo! in the United States. A crawler-type search engine was used as well, and as the popularity of the crawler-type search engine gradually increased, after October 3, 2005, Yahoo! Japan began utilizing only the crawler-type engine. On June 29, 2017, Yahoo! Japan announced that the directory-based search engine "Yahoo! Category", which had been in operation since its establishment, would be abolished on March 29, 2018.[3]

As a crawler-type search engine, Yahoo! Japan initially used technology from the Japanese company Goo, which used Google's technology. The company later switched to using Yahoo Search Technology (YST), developed by Yahoo! in the US. In addition to serving as a standard search engine, Yahoo! Japan partnered with Twitter to provide real-time search for tweets.[4] It also receives data feeds from partner companies; Cookpad and Naver information is displayed in search results. Yahoo! Search Custom Search was discontinued on March 31, 2019.[5]

History
Yahoo! and SoftBank formed Yahoo! Japan in January 1996 to establish the first web portal in Japan. Yahoo! Japan went live on April 1, 1996.[6] Yahoo! Japan was listed on JASDAQ in November 1997. In January 2000, it became the first stock in Japanese history to trade for more than ¥100 million per share. The company was listed on the Tokyo Stock Exchange in October 2003 and became part of the Nikkei 225 stock market index in 2005.

Yahoo! Japan acquired the naming rights for the Fukuoka Dome in 2005, renaming the dome as the "Fukuoka Yahoo! Japan Dome". The "Yahoo Dome" is the home field for the Fukuoka SoftBank Hawks, a professional baseball team majority owned by SoftBank.

Since 2010, Yahoo! Japan's search engine has been based on Google's search technology. In exchange, Google receives user activity data from Yahoo! Japan's various products.[7]

In 2017, Verizon Communications purchased the core internet business of United States-based Yahoo!, and merged it with AOL into Oath, Inc.; Yahoo! Japan was not affected.

Yahoo Japan's services are not available in the European Economic Area and the UK since 6 April 2022, due to "excessive regulatory burden".[8][9]

Yahoo Japan, whose partnership with Google on search engine technology is set to expire in 2025, is considering switching to South Korean company Naver's search engine technology.[10]